Changed the default weights. Weight dialog box bug fix.
Changed the default weights of the algorithm.
Changed the dialog box on the weights page so that the user cannot click outside of the box. Previously if they clicked outside of the box it would cause an error and prevent them from accessing that weight in the future.
Changed some of the formatting for the help page.