Location page addition
Added the first batch of functionality to the location page. There is now a button that, when pressed, will prompt the user to pick a file folder, same as the buttons on the upload page, then scan the data for locations visited in the past two weeks. Only works for Facebook at the moment, but the framework is in place so that incorporating Instagram shouldn't take much time.