Changed min version. Changed button colors and space formatting.
Had an issue related to the ego network results overflowing so I changed the space formatting very slightly to fix that. Changed the buttons on the ego network page to be Raised Buttons to give more emphasis as well as testing giving the buttons color. Changed minSdkVersion in build.gradle to 23. That should force it to use AndroidX?